refactor: fetchAllEntity 에서 fetchAllModel로 수정
- 예를들어 Realm으로 바뀐다면 Entity는 coredata에 종속적이니까 Model을 주는 것이 맞다고 생각함
- Model의 값이 double에서 Int로 바뀔부분 수정

+++ TempusTests/BlockUseCaseTests/BlockStartUseCaseTests.swift
@@ -47,10 +47,10 @@ class BlockStartUseCaseTests: XCTestCase {
         wait(for: [expectation], timeout: 1.0)
     }
     
-    final private func generateSchedule(divideCount: Double) -> [Date] {
+    final private func generateSchedule(divideCount: Int) -> [Date] {
         let calendar = Calendar.current
         let now = Date()
-        let interval = 24 / divideCount
+        let interval = Double(24 / divideCount)
         let oneDaySecond = 24.0 * 60.0 * 60.0
         let oneHourSecond = 60.0 * 60.0
